feat: adding google pay on pdp (#1198)
