+++ src/cartridges/app_adyen_SFRA/cartridge/templates/default/cart/checkoutButtons.isml
@@ -86,7 +86,7 @@
                 <div data-method="applepay" class="expressComponent applepay" style="padding:0"></div>
                 <div data-method="amazonpay" class="expressComponent" id="amazonpay-container"></div>
                 <div data-method="paypal" class="expressComponent" id="paypal-container" style="padding:0"></div>
-                <div data-method="googlepay" class="expressComponent" id="googlepay-container" style="padding:0"></div>
+                <div data-method="googlepay" class="expressComponent googlepay" style="padding:0"></div>
             </div>
             <script src="${URLUtils.httpsStatic('/js/expressPaymentMethodsVisibility.js')}" type="text/javascript"></script>
         </isif>
